# contracts/openapi/sirta-v1.yaml
openapi: 3.1.0
info:
  title: SIRTA Municipal API
  version: 0.3.0
  description: >
    Active OpenAPI source of truth for SIRTA. implementation_version 0.3.42
    serves official PUBLIC_OPEN Gold pending human validation on the 0.3.0 specification.
    Roadmap stage OFFICIAL_DATA_WAVE_1_COMPLETE_ROADMAP_IN_PROGRESS.
    ISR and transfer writes remain specified and are not served until later local slices.
    Sensitive operations require OIDC plus tenant, territory and purpose context.
servers:
  - url: http://localhost:8080
security:
  - oidc: []
paths:
...
  /v1/indicators/source-enrichment:
    get:
      operationId: getSourceEnrichment
      description: Published official Gold enrichment. Public sources never create tax credits.
      parameters:
        - name: sourceId
          in: query
          required: true
          schema: {type: string}
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Enrichment counts; createsTaxCredit=false}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/active-debt:
    get:
      operationId: getActiveDebtPanel
      description: >
        Official empty active-debt panel. inscriptionEnabled=false;
        no legal inscription until municipal authorization and G0/G5/G6.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Empty active-debt panel; inscription disabled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: inscribeActiveDebt
      description: Always rejected (409). Technical shell never inscribes active debt.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '409': {description: Inscription disabled; no certidão}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/transfer-reconciliation:
    get:
      operationId: getTransferReconciliation
      description: >
        Empty transfer reconciliation shell. Compares expected versus received only
        when both exist. Differences are occurrences, never tax credits. G7 official blocked.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200': {description: Empty reconciliation panel; createsTaxCredit=false}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: commandTransferReconciliation
      description: Always rejected (409). No automatic reconciliation or credit creation.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '409': {description: Reconciliation commands disabled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/data-sources:
    get:
      operationId: listDataSources
      description: Official master source registry. Public sources never create tax credits.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Catalog items with ingestAllowed and createsTaxCredit=false}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/data-sources/{sourceId}/dry-run:
    post:
      operationId: dryRunDataSource
      description: Dry-run of a cataloged source. Official sources report the planned official URL.
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: {type: string}
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Dry-run accepted for an allowed catalog source}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/data-sources/{sourceId}/ingest:
    post:
      operationId: ingestCatalogSource
      description: >
        Catalog-driven official ingest for PUBLIC_OPEN sources. Preserves landing checksum,
        quarantines invalid rows and never creates a tax credit. Synthetic fixtures are test-only.
      parameters:
        - name: sourceId
          in: path
          required: true
          schema: {type: string}
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Ingest accepted or replayed; taxCreditCreated=false}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/indicators/official-gold:
    get:
      operationId: getOfficialGold
      description: >
        Official Gold indicators derived from Silver. Homologation remains
        REAL_OFFICIAL_DATA_PENDING_HUMAN_VALIDATION. Empty sources are listed, never simulated.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Official Gold items or empty sources with reason}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/indicators/official-gold/lines:
    get:
      operationId: getOfficialGoldLines
      description: >
        Per-row Gold lineage to Silver, Bronze checksum, landing manifest and official URL.
        Incomplete lineage is rejected.
      parameters:
        - name: sourceId
          in: query
          schema: {type: string}
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Gold lines with complete lineage}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/dashboards:
    get:
      operationId: listDashboards
      description: Fifteen official dashboards. Empty modules stay empty. No synthetic fill. Commands disabled.
      parameters:
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200': {description: Dashboard catalog with official Gold subsets}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/dashboards/{dashboardId}:
    get:
      operationId: getDashboard
      description: >
        One official dashboard with optional kpis/charts from PUBLIC_OPEN Gold.
        ROI is never calculated without eligible received values. Charts only use
        REFERENCE_QUANTITY or TRANSFER_AMOUNT_AS_PUBLISHED. Not tax credit.
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema: {type: string}
        - {$ref: '#/components/parameters/TerritoryId'}
        - {$ref: '#/components/parameters/PurposeId'}
      responses:
        '200':
          description: Dashboard view with Gold items, kpis and charts
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DashboardView'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8081/realms/sirta/.well-known/openid-configuration
  parameters:
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    PageSize:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    TerritoryId:
      name: X-Territory-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    PurposeId:
      name: X-Purpose-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
  schemas:
...
    DashboardChartPoint:
      type: object
      required: [x, y]
      properties:
        x: {type: string}
        y: {type: number}
    DashboardChartSeries:
      type: object
      required: [name, points]
      properties:
        name: {type: string}
        points:
          type: array
          items: {$ref: '#/components/schemas/DashboardChartPoint'}
    DashboardChart:
      type: object
      required: [id, title, type, unit, valueKind, series, evidenceIds]
      properties:
        id: {type: string}
        title: {type: string}
        type: {type: string, enum: [bar, line, pie]}
        unit: {type: string}
        valueKind: {type: string}
        series:
          type: array
          items: {$ref: '#/components/schemas/DashboardChartSeries'}
        evidenceIds:
          type: array
          items: {type: string}
    DashboardKpi:
      type: object
      required: [id, label, value, unit, valueKind, sourceId]
      properties:
        id: {type: string}
        label: {type: string}
        value: {type: number}
        unit: {type: string}
        valueKind: {type: string}
        sourceId: {type: string}
        evidenceId: {type: [string, 'null']}
    DashboardView:
      type: object
      required:
        - id
        - title
        - banner
        - homologationStatus
        - createsTaxCredit
        - commandsDisabled
        - published
        - items
        - emptySources
        - roiCalculated
        - taxPotentialAsCredit
        - kpis
        - charts
      properties:
        id: {type: string}
        path: {type: string}
        title: {type: string}
        banner: {type: string}
        homologationStatus:
          type: string
          enum:
            - REAL_OFFICIAL_DATA_HUMAN_VALIDATED_REFERENCE_ONLY
            - REAL_OFFICIAL_DATA_PENDING_HUMAN_VALIDATION
        createsTaxCredit: {type: boolean}
        commandsDisabled: {type: boolean}
        emptyReason: {type: string}
        published: {type: boolean}
        items: {type: array, items: {type: object, additionalProperties: true}}
        emptySources: {type: array, items: {type: object, additionalProperties: true}}
        roiCalculated: {type: boolean}
        taxPotentialAsCredit: {type: boolean}
        kpis:
          type: array
          items: {$ref: '#/components/schemas/DashboardKpi'}
        charts:
          type: array
          items: {$ref: '#/components/schemas/DashboardChart'}
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        traceId: {type: string}
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: Missing role, tenant, territory or valid access purpose
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found or not visible in the current context
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...